Refactor Alert Message

+++ Core/Modules/TeamModule/Remove/RemoveTeamHandler.cs
@@ -11,10 +11,12 @@ namespace Core.Modules.TeamModule.Remove
     public class RemoveTeamHandler : IRequestHandler<RemoveTeamCommand, ActionResponse>
     {
         private readonly ITeamRepository _teamRepository;
+        private readonly IGroupDetailsRepository _groupDetailsRepository;
 
-        public RemoveTeamHandler(ITeamRepository teamRepository)
+        public RemoveTeamHandler(ITeamRepository teamRepository, IGroupDetailsRepository groupDetailsRepository)
         {
             _teamRepository = teamRepository;
+            _groupDetailsRepository = groupDetailsRepository;
         }
 
         public async Task<ActionResponse> Handle(RemoveTeamCommand request, CancellationToken cancellationToken)
@@ -23,7 +25,11 @@ namespace Core.Modules.TeamModule.Remove
             if (team == null)
                 return new ActionResponse { IsSuccess = false, Title = "Error", Message = "The team does not exist", State = State.error };
 
-            if(!await _teamRepository.DeleteTeamAsync(team))
+            if(await _groupDetailsRepository.GetGroupDetailsByTeamAsync(team.Id) != null)
+                return new ActionResponse { IsSuccess = false, Title = "Error", Message = "The team is currently in a tournament", State = State.error };
+
+
+            if (!await _teamRepository.DeleteTeamAsync(team))
                 return new ActionResponse { IsSuccess = false, Title = "Error", Message = "Something has gone wrong", State = State.error };
                 
             return new ActionResponse { IsSuccess = false, Title = "Deleted!", Message = $"Team {team.Name} has been deleted!", State = State.success };
